+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.0.3
+
+* Add application/toml mimeType lookup by extension.
+* Add font/woff2 lookup by extension and header bytes.
+* Add image/avif mimeType lookup by extension.
+* Add image/heic mimeType lookup by extension.
+* Add image/heif mimeType lookup by extension.
+
+# 1.0.2
+
+* Add audio/x-aiff mimeType lookup by header bytes.
+* Add audio/x-flac mimeType lookup by header bytes.
+* Add audio/x-wav mimeType lookup by header bytes.
+* Add audio/mp4 mimeType lookup by file path.
+
 # 1.0.1
 
 * Add image/webp mimeType lookup by header bytes.